refactor(web): add useProposalChanges hook
Based on a query and as a replacement of client.software.onSelectedPatternsChange.
